Slimmed down runtime crate that only wraps the runtime core crate
